requirements.txt
asgiref==3.2.3
backcall==0.1.0
colorama==0.4.3
decorator==4.4.1
Django==2.2.6
django-cors-headers==3.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.2
reverse==0.1.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
